requirements.txt
flask
mlflow
sentence-transformers
datasets
pandas
pyspark
faiss-cpu
numpy